Moved in-game settings menu to main container Jframe. Renamed it to FILE. AB#1322
