+++ src/main/java/com/game/templejog/gui/top/TopHUD.java
@@ -16,13 +16,6 @@ public class TopHUD {
     private JProgressBar currentHealthBar;
     private JLabel hoursPlayed;
     private final JButton helpButton;
-    //    private final JButton quitButton;
-    private JMenuBar pauseBar;
-    private JMenu pauseMenu;
-    private JMenuItem saveItem;
-    private JMenuItem quitItem;
-    private JMenuItem loadItem;
-    private JMenuItem settingsItem;
 
     public TopHUD(Game game) {
         this.topHUDJPanel = new JPanel();
@@ -37,47 +30,6 @@ public class TopHUD {
                 HelpMenu.setUpHelpGUI(this.game);
             }
         });
-
-
-        this.pauseBar = new JMenuBar();
-        this.pauseMenu = new JMenu("FILE");
-        saveItem = new JMenuItem("SAVE");
-        saveItem.addActionListener(e -> {
-            if (e.getSource() == this.saveItem) {
-                saveGame(game);
-            }
-        });
-
-
-        loadItem = new JMenuItem("LOAD");
-        loadItem.addActionListener(e -> {
-            if (e.getSource() == this.loadItem) {
-                loadGame(game);
-            }
-        });
-
-        settingsItem = new JMenuItem("SETTINGS");
-        settingsItem.addActionListener(e -> {
-            if (SettingsMenu.activeWindow) {
-                SettingsMenu.settingsFrame.dispose();
-                SettingsMenu.activeWindow = false;
-            } else {
-                SettingsMenu.settingsMenuDisplay(game);
-            }
-        });
-
-        quitItem = new JMenuItem("QUIT GAME");
-        quitItem.addActionListener(e -> {
-            QuitMenu.setUpInGameQuitOptions(MainContainer.getMainContainer());
-        });
-
-        pauseMenu.add(saveItem);
-        pauseMenu.add(loadItem);
-        pauseMenu.add(settingsItem);
-        pauseMenu.add(quitItem);
-
-        pauseBar.add(pauseMenu);
-
     }
 
 
@@ -95,35 +47,8 @@ public class TopHUD {
         topHUDJPanel.add(this.currentHealthBar);
         topHUDJPanel.add(this.hoursPlayed);
         topHUDJPanel.add(this.helpButton);
-//        topHUDJPanel.add(this.quitButton);
-        topHUDJPanel.add(this.pauseBar);
         topHUDJPanel.setVisible(true);
         return topHUDJPanel;
-    }
-
-    private static void saveGame(Game game) {
-        try {
-            FileOutputStream fos = new FileOutputStream("TempleJog.sav");
-            ObjectOutput oos = new ObjectOutputStream(fos);
-            oos.writeObject(game);
-            oos.flush();
-            oos.close();
-            System.out.println("Game saved\n");
-        } catch (Exception e) {
-            System.out.println("Serialization error: " + e.getClass() + ": " + e.getMessage());
-        }
-    }
-
-    private static void loadGame(Game game) {
-        try {
-            FileInputStream fis = new FileInputStream("TempleJog.sav");
-            ObjectInputStream ois = new ObjectInputStream(fis);
-            game = (Game) ois.readObject();
-            ois.close();
-            System.out.println("\n Game loaded");
-        } catch (Exception e) {
-            System.out.println("Deserialization error: " + e.getClass() + ": " + e.getMessage());
-        }
     }
 
     public Game getGame() {
@@ -134,27 +59,8 @@ public class TopHUD {
         this.game = game;
     }
 
-    public JLabel getCurrentLocation() {
-        return currentLocation;
-    }
-
-    public void setCurrentLocation(JLabel currentLocation) {
-        this.currentLocation = currentLocation;
-    }
-
-    public JLabel getCurrentHealth() {
-        return currentHealth;
-    }
-
-    public void setCurrentHealth(JLabel currentHealth) {
-        this.currentHealth = currentHealth;
-    }
-
     public JPanel getTopHUDJPanel() {
         return topHUDJPanel;
     }
 
-    public void setTopHUDJPanel(JPanel topHUDJPanel) {
-        this.topHUDJPanel = topHUDJPanel;
-    }
 }
